feat: add routing, navigation panel, and main page - configure React Router in App so it only renders routes and pages - introduce NavMenu component with navigation links - create MainPage component and move SearchBar and ResultsField logic into it - add basic styles for navigation bar and container layout - add folder utils amd util navLinkClass.ts
