+++ rs-react-app/src/App.tsx
@@ -1,29 +1,14 @@
-import { useEffect, useState } from 'react';
-import { ErrorBoundary } from './components/pages/SearchResults/ErrorBoundary.tsx';
-import SearchBarFn from './components/pages/SearchResults/SearchBarFn.tsx';
-import ResultsFieldFn from './components/pages/SearchResults/ResultsFieldFn.tsx';
+import NavMenu from './components/NavMenu.tsx';
+import { Route, Routes } from 'react-router-dom';
+import SearchPage from './components/pages/SearchResults';
 
 function App() {
-  const [search, setSearch] = useState('');
-  const [triggerSearch, setTriggerSearch] = useState(false);
-  useEffect(() => {
-    const saveSearch = localStorage.getItem('search');
-    if (saveSearch) {
-      setSearch(saveSearch);
-    }
-  }, []);
-
-  const handleSearch = (search: string) => {
-    setSearch(search);
-    setTriggerSearch(!triggerSearch);
-  };
-
   return (
     <div className="container">
-      <SearchBarFn search={search} onSearch={handleSearch} />
-      <ErrorBoundary>
-        <ResultsFieldFn search={search} triggerSearch={triggerSearch} />
-      </ErrorBoundary>
+      <NavMenu />
+      <Routes>
+        <Route path="/" element={<SearchPage />}></Route>
+      </Routes>
     </div>
   );
 }
