+++ rs-react-app/src/components/NavMenu.scss
@@ -1,0 +1,13 @@
+.navbar {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+
+  gap: 2em;
+  .pending{
+    color: red;
+  }
+  .active{
+    color: blue;
+  }
+}
