+++ rs-react-app/src/components/utils/navLinkClass.ts
@@ -1,0 +1,11 @@
+export function navLinkClass({
+  isActive,
+  isPending,
+}: {
+  isActive: boolean;
+  isPending: boolean;
+}) {
+  if (isPending) return 'pending';
+  if (isActive) return 'active';
+  return '';
+}
